Fix various errors in the gcc compile:  added socket_handle_valid() since socket_handle_t is unsigned (so comparison with <0 doesn't work). Removed register, added missing functions in template (modem) and added code to read device id from file.

+++ services/src/debug.c
@@ -14,7 +14,6 @@
 #include "debug.h"
 #include "timer_hal.h"
 
-
 uint32_t log_level_at_run_time = LOG_LEVEL_AT_RUN_TIME;
 
 void log_print_(int level, int line, const char *func, const char *file, const char *msg, ...)
@@ -31,7 +30,7 @@ void log_print_(int level, int line, const char *func, const char *file, const c
         va_list args;
         va_start(args, msg);
         file = file ? strrchr(file,'/') + 1 : "";
-        int trunc = snprintf(_buffer, arraySize(_buffer), "%010lu:<%s> %s %s(%d):", HAL_Timer_Get_Milli_Seconds(), levels[level], func, file, line);
+        int trunc = snprintf(_buffer, arraySize(_buffer), "%010u:<%s> %s %s(%d):", (unsigned)HAL_Timer_Get_Milli_Seconds(), levels[level], func, file, line);
         if (debug_output_)
         {
             debug_output_(_buffer);
